Frontend: split in files and add functions to redirect link to connect to intra
