+++ data/frontend/package.json
@@ -19,6 +19,8 @@
     "http-proxy-middleware": "^2.0.6",
     "react": "^18.2.0",
     "react-dom": "^18.2.0",
+    "react-router": "^6.14.2",
+    "react-router-dom": "^6.14.2",
     "react-scripts": "^5.0.1",
     "socket.io": "^4.6.1",
     "socket.io-client": "^4.6.1",
