+++ data/frontend/src/components/Home/Home.tsx
@@ -1,0 +1,9 @@
+import Navigation from "../Navigation/Navigation";
+
+function Home() {
+    return (
+      <Navigation />
+    );
+}
+
+export default Home;
